move to config api
